null checks in commands

+++ AskerTracker.Application/Features/Fees/Commands/DeleteFees/DeleteFeeCommandHandler.cs
@@ -1,4 +1,5 @@
 using AskerTracker.Application.Contracts.Persistence;
+using AskerTracker.Application.Exceptions;
 using AskerTracker.Domain.Entities;
 using AutoMapper;
 using MediatR;
@@ -19,6 +20,9 @@ public class DeleteFeeCommandHandler : IRequestHandler<DeleteFeeCommand>
     public async Task<Unit> Handle(DeleteFeeCommand request, CancellationToken cancellationToken)
     {
         var feeToDelete = await _feeRepository.GetByIdAsync(request.Id);
+
+        if (feeToDelete == null)
+            throw new NotFoundException(nameof(MembershipFee), request.Id);
 
         await _feeRepository.DeleteAsync(feeToDelete);
